+++ packages/models/src/lib/implementation/schemas.ts
@@ -119,7 +119,7 @@ export const positiveIntSchema = z.number().int().positive();
 
 export const nonnegativeIntSchema = z.number().int().nonnegative();
 
-export const nonnegativeFloatSchema = z.number().nonnegative();
+export const nonnegativeNumberSchema = z.number().nonnegative();
 
 export function packageVersionSchema<TRequired extends boolean>(options?: {
   versionDescription?: string;
@@ -142,11 +142,9 @@ export function packageVersionSchema<TRequired extends boolean>(options?: {
 }
 
 /** Schema for a weight */
-export const weightSchema = z
-  .union([nonnegativeIntSchema, nonnegativeFloatSchema])
-  .describe(
-    'Coefficient for the given score (use weight 0 if only for display)',
-  );
+export const weightSchema = nonnegativeNumberSchema.describe(
+  'Coefficient for the given score (use weight 0 if only for display)',
+);
 
 export function weightedRefSchema(
   description: string,