+++ packages/plugin-js-packages/src/lib/runner/runner.integration.test.ts
@@ -12,6 +12,8 @@ describe('createRunnerConfig', () => {
       packageManager: 'npm',
       checks: ['audit'],
       auditLevelMapping: defaultAuditLevelMapping,
+      dependencyGroups: ['prod', 'dev'],
+      packageJsonPath: 'package.json',
     });
     expect(runnerConfig).toStrictEqual<RunnerConfig>({
       command: 'node',
@@ -25,7 +27,9 @@ describe('createRunnerConfig', () => {
     const pluginConfig: FinalJSPackagesPluginConfig = {
       packageManager: 'yarn-classic',
       checks: ['outdated'],
+      dependencyGroups: ['prod', 'dev'],
       auditLevelMapping: { ...defaultAuditLevelMapping, moderate: 'error' },
+      packageJsonPath: 'package.json',
     };
     await createRunnerConfig('executeRunner.ts', pluginConfig);
     const config = await readJsonFile<FinalJSPackagesPluginConfig>(